refactor: move media metadata utilities to media.ts
